requirements.txt
crewai
fastapi
uvicorn[standard]
pydantic
PyGithub
gitpython
python-dotenv
requests